requirements.txt
torch
torchvision
transformers
Pillow
tqdm
numpy
scikit-learn
scipy
